Move utility components to /src/components folder
