+++ src/components/BoldedText/index.js
@@ -1,0 +1,3 @@
+import BoldedText from './BoldedText';
+
+export default BoldedText;
