+++ src/components/HeadModifier/index.js
@@ -1,0 +1,3 @@
+import HeadModifier from './HeadModifier';
+
+export default HeadModifier;
