+++ src/layouts/components/PageHandler/PageHandler.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { intlShape } from 'react-intl';
 import { uppercaseFirst } from '../../../utils';
-import HeadModifier from '../../../utils/headModifier';
+import HeadModifier from '../../../components/HeadModifier';
 
 class PageHandler extends React.Component {
   componentDidMount() {
